+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/crypto/model/RoomKeyBundleContent.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Matrix.org Foundation C.I.C.
+ * Copyright 2026 The Matrix.org Foundation C.I.C.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.matrix.android.sdk.api.session.crypto.model
 
-package org.matrix.android.sdk.internal.crypto.model
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
 
-data class SessionInfo(
-        val sessionId: String,
-        val senderKey: String
+/**
+ * Content of an MSC4268 `m.room_key_bundle` to-device message: where to find the key bundle and how to decrypt it.
+ */
+@JsonClass(generateAdapter = true)
+data class RoomKeyBundleContent(
+        @Json(name = "room_id")
+        val roomId: String? = null,
+
+        @Json(name = "file")
+        val file: EncryptedFileInfo? = null,
 )